rename package crm to resourcemanager
